split catergories and todo_categories routes & models
